+++ rconweb/api/history.py
@@ -7,22 +7,15 @@ from django.http import JsonResponse
 from django.views.decorators.csrf import csrf_exempt
 
 from rcon.utils import MapsHistory
-from rcon.recorded_commands import RecordedRcon
+
 from rcon.commands import CommandFailedError
-from rcon.steam_utils import get_steam_profile
-from rcon.settings import SERVER_INFO
 from rcon.player_history import (
     get_players_by_appearance,
-    add_player_to_blacklist,
-    remove_player_from_blacklist,
     get_player_profile,
     get_player_profile_by_id,
     add_flag_to_player,
     remove_flag,
 )
-from rcon.game_logs import LogLoop
-from rcon.user_config import AutoBroadcasts, InvalidConfigurationError, StandardMessages
-from rcon.cache_utils import RedisCached, get_redis_pool
 from .auth import login_required, api_response
 from .utils import _get_data
 from rcon.discord import send_to_discord_audit
